20230914
1.資料庫更新
2.新增-> GetTwseStockDayAndInsertToDbAsync

+++ JackStockApi/Domain/StockDayHistory.cs
@@ -49,7 +49,7 @@ namespace JackStockApi.Domain
         /// <summary>
         /// 成交筆數
         /// </summary>
-        public int Transaction { get; set; }
+        public long Transaction { get; set; }
 
 
         public virtual Stock Stock { get; set; }
